[test.math] fix OtherMath, AbsoluteValue Test
This did not compile for me with clang 3.8.1-12~bpo8+1 using gcc
4.9.2-10 stdlib. Any objections against this change, @dune-xt-common-devs?
Was there any reason to use int instead of double in the first place?

+++ dune/xt/common/test/math.cc
@@ -117,8 +117,8 @@ TEST(OtherMath, AbsoluteValue)
   EXPECT_EQ(abs(-1l), 1l);
   EXPECT_EQ(Dune::XT::Common::abs(0u), 0u);
   EXPECT_EQ(abs(0), 0);
-  EXPECT_EQ(abs(std::complex<int>(0)), 0);
-  EXPECT_EQ(abs(std::complex<int>(-1)), 1);
+  EXPECT_EQ(abs(std::complex<double>(0)), 0);
+  EXPECT_EQ(abs(std::complex<double>(-1)), 1);
 }
 
 TEST(OtherMath, FloatCmp)
